Restored code to create a gradient entropy image
This was an earlier mistake, wherein I applied the entropy to the
gradient image instead of the histogram image. It’s now restored as a
separate function. Note that the naming conventions have changed to
something more sensible: _grad_ent for the new gradient entropy image,
and _hist_ent for the existing histogram entropy image.